# backend/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Festivals API",
    "description": "API for managing festivals, cashless payments, tickets, and more. This API provides comprehensive endpoints for festival management including wallet operations, NFC payments, ticket scanning, and administrative functions.",
    "termsOfService": "https://festivals.io/terms",
    "contact": {
      "name": "API Support",
      "url": "https://festivals.io/support",
      "email": "support@festivals.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'"
    },
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "X-API-Key",
      "in": "header",
      "description": "API Key for external integrations"
    }
  },
...
  "paths": {
...
    "/festivals/{id}": {
      "get": {
        "tags": ["festivals"],
        "summary": "Get festival by ID",
        "description": "Get detailed information about a specific festival",
        "operationId": "getFestival",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "type": "string", "format": "uuid", "description": "Festival ID"}
        ],
        "responses": {
          "200": {"description": "Festival details", "schema": {"$ref": "#/definitions/FestivalResponse"}},
          "400": {"description": "Invalid ID", "schema": {"$ref": "#/definitions/ErrorResponse"}},
          "404": {"description": "Festival not found", "schema": {"$ref": "#/definitions/ErrorResponse"}}
        }
      },
      "patch": {
        "tags": ["festivals"],
        "summary": "Update festival",
        "description": "Update an existing festival with partial data",
        "operationId": "updateFestival",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "type": "string", "format": "uuid"},
          {"name": "body", "in": "body", "required": true, "schema": {"$ref": "#/definitions/UpdateFestivalRequest"}}
        ],
        "responses": {
          "200": {"description": "Updated festival", "schema": {"$ref": "#/definitions/FestivalResponse"}},
          "400": {"description": "Invalid request", "schema": {"$ref": "#/definitions/ErrorResponse"}},
          "404": {"description": "Festival not found", "schema": {"$ref": "#/definitions/ErrorResponse"}}
        }
      },
      "delete": {
        "tags": ["festivals"],
        "summary": "Delete festival",
        "description": "Permanently delete a festival",
        "operationId": "deleteFestival",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "type": "string", "format": "uuid"}
        ],
        "responses": {
          "204": {"description": "Festival deleted"},
          "404": {"description": "Festival not found", "schema": {"$ref": "#/definitions/ErrorResponse"}}
        }
      }
    },
...
  },
  "definitions": {
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {"$ref": "#/definitions/ErrorDetail"}
      }
    },
    "ErrorDetail": {
      "type": "object",
      "properties": {
        "code": {"type": "string"},
        "message": {"type": "string"},
        "details": {"type": "object"}
      }
    },
...
    "UpdateFestivalRequest": {
      "type": "object",
      "properties": {
        "name": {"type": "string"},
        "description": {"type": "string"},
        "startDate": {"type": "string", "format": "date-time"},
        "endDate": {"type": "string", "format": "date-time"},
        "location": {"type": "string"}
      }
    },
    "FestivalResponse": {
      "type": "object",
      "properties": {
        "id": {"type": "string", "format": "uuid"},
        "name": {"type": "string"},
        "description": {"type": "string"},
        "startDate": {"type": "string", "format": "date-time"},
        "endDate": {"type": "string", "format": "date-time"},
        "location": {"type": "string"},
        "status": {"type": "string", "enum": ["draft", "active", "archived"]},
        "currencyName": {"type": "string"},
        "exchangeRate": {"type": "number"},
        "createdAt": {"type": "string", "format": "date-time"},
        "updatedAt": {"type": "string", "format": "date-time"}
      }
    },
...
  }
}